Change Game.RunTime to a long to prevent overflow.

+++ OpenRA.Game/Network/OrderManager.cs
@@ -39,7 +39,7 @@ namespace OpenRA.Network
 		public int LocalFrameNumber;
 		public int FramesAhead = 0;
 
-		public int LastTickTime = Game.RunTime;
+		public long LastTickTime = Game.RunTime;
 
 		public bool GameStarted { get { return NetFrameNumber != 0; } }
 		public IConnection Connection { get; private set; }
